pyproject.toml
[tool.poetry]
name = "seasonwatch"
version = "0.1.0"
description = "A script to let you know when there are new seasons of your TV shows"
authors = ["Simon <gevhaz@tutanota.com>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.9"
IMDbPY = "^2021.4.18"
PyYAML = "^6.0"
colorama = "^0.4.5"
python-dateutil = "^2.8.2"
PyGObject = "^3.42.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.961"
types-PyYAML = "^6.0.8"
types-python-dateutil = "^2.8.18"

[tool.poetry.scripts]
seasonwatch = 'seasonwatch.app:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = [
    "imdb",
    "imdb.parser.http",
    "imdb.parser.s3",
    "imdb.parser.sql",
    "colorama",
    "gi",
    "gi.repository"
]
ignore_missing_imports = true

[tool.isort]
profile = "black"
